lib.align.AlignedFace - remove Extract class
