+++ lib/align/__init__.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 """ Package for handling alignments files, detected faces and aligned faces along with their
 associated objects. """
-from .aligned_face import Extract, get_matrix_scaling, PoseEstimate  # noqa
+from .aligned_face import AlignedFace, get_matrix_scaling, PoseEstimate, transform_image  # noqa
 from .alignments import Alignments, Thumbnails  # noqa
-from .detected_face import AlignedFace, BlurMask, DetectedFace, Mask  # noqa
+from .detected_face import BlurMask, DetectedFace, Mask  # noqa
